Shifted generic functions from RuneLiteBot to Bot
- Removed some unused images
- Fixed the odd bug
