+++ src/model/bot.py
@@ -3,15 +3,21 @@ A Bot is a base class for bot script models. It is abstract and cannot be instan
 pre-implemented and can be used by subclasses, or called by the controller. Code in this class should not be modified.
 '''
 from abc import ABC, abstractmethod
+from deprecated import deprecated
 from enum import Enum
 from model.window import Window
 from threading import Thread
+from utilities.geometry import Point, Rectangle
 from utilities.mouse_utils import MouseUtils
 from utilities.options_builder import OptionsBuilder
 import customtkinter
 import keyboard
+import numpy as np
+import pyautogui as pag
 import pygetwindow
+import pytweening
 import time
+import utilities.bot_cv as bcv
 import warnings
 warnings.filterwarnings("ignore", category=UserWarning)
 
@@ -33,6 +39,9 @@ class Bot(ABC):
     status = BotStatus.STOPPED
     thread: Thread = None
     win: Window = None
+
+    GREEN = [0, 255, 0]
+    RED = [255, 0, 0]
 
     # ---- Abstract Functions ----
     @abstractmethod
@@ -242,3 +251,158 @@ class Bot(ABC):
         Requests the controller to tell the UI to clear the log.
         '''
         self.controller.clear_log()
+
+    # --- Misc Utility Functions
+    def drop_inventory(self, skip_rows: int = 0, skip_slots: list[int] = None) -> None:
+        '''
+        Drops all items in the inventory.
+        Args:
+            skip_rows: The number of rows to skip before dropping.
+            skip_slots: The indices of slots to avoid dropping.
+        '''
+        self.log_msg("Dropping inventory...")
+        # Determine slots to skip
+        if skip_slots is None:
+            skip_slots = []
+        if skip_rows > 0:
+            row_skip = list(range(skip_rows*4))
+            skip_slots = np.unique(row_skip + skip_slots)
+        # Start dropping
+        pag.keyDown("shift")
+        for i, slot in enumerate(self.win.inventory_slots):
+            if not self.status_check_passed():
+                pag.keyUp("shift")
+                return
+            if i in skip_slots:
+                continue
+            p = slot.random_point()
+            self.mouse.move_to((p[0], p[1]),
+                                mouseSpeed='fastest',
+                                knotsCount=1,
+                                offsetBoundaryY=40,
+                                offsetBoundaryX=40,
+                                tween=pytweening.easeInOutQuad)
+            pag.click()
+        pag.keyUp("shift")
+
+    def friends_nearby(self) -> bool:
+        '''
+        Checks the minimap for green dots to indicate friends nearby.
+        Returns:
+            True if friends are nearby, False otherwise.
+        '''
+        # screenshot minimap
+        minimap = bcv.screenshot(self.win.minimap)
+        bcv.save_image("minimap.png", minimap)
+        only_friends = bcv.isolate_colors(minimap, [self.GREEN])
+        bcv.save_image("minimap_friends.png", only_friends)
+        mean = only_friends.mean(axis=(0, 1))
+        return mean != 0.0
+
+    def logout(self):
+        '''
+        Logs player out.
+        '''
+        self.log_msg("Logging out...")
+        self.mouse.move_to(self.win.cp_tabs[10].random_point())
+        pag.click()
+        time.sleep(1)
+        self.mouse.move_rel(0, -53, 5, 5)  # Logout button
+        pag.click()
+
+    # --- Player Status Functions ---
+    def has_hp_bar(self) -> bool:
+        '''
+        Returns whether the player has an HP bar above their head. Useful alternative to using OCR to check if the
+        player is in combat. This function only works when the game camera is all the way up.
+        '''
+        # Position of character relative to the screen
+        char_pos = self.win.game_view.get_center()
+
+        # Make a rectangle around the character
+        offset = 30
+        char_rect = Rectangle.from_points(Point(char_pos.x - offset, char_pos.y - offset*2),
+                                          Point(char_pos.x + offset, char_pos.y))
+        # Take a screenshot of rect
+        char_screenshot = bcv.screenshot(char_rect)
+        # Isolate HP bars in that rectangle
+        hp_bars = bcv.isolate_colors(char_screenshot, [self.RED, self.GREEN])
+        # If there are any HP bars, return True
+        return hp_bars.mean(axis=(0, 1)) != 0.0
+    
+    @deprecated(reason="The OCR this function uses may be innacurate. Consider using an API function, or check colors on the win.hp_bar.")
+    def get_hp(self) -> int:
+        """
+        Gets the HP value of the player.
+        Returns:
+            The HP of the player, or None if not found.
+        """
+        res = bcv.get_numbers_in_rect(self.win.hp_orb_text, True)
+        print(res)
+        return None if res is None else res[0]
+
+    @deprecated(reason="The OCR this function uses may be innacurate. Consider using an API function, or check colors on the win.prayer_bar.")
+    def get_prayer(self) -> int:
+        """
+        Gets the prayer value of the player.
+        Returns:
+            The prayer value of the player, or None if not found.
+        """
+        res = bcv.get_numbers_in_rect(self.win.prayer_orb_text, True)
+        print(res)
+        return None if res is None else res[0]
+
+    # --- Client Settings ---
+    # TODO: Add anti-ban functions that move camera around randomly
+    def move_camera_up(self):
+        '''
+        Moves the camera up.
+        '''
+        # Position the mouse somewhere on the game view
+        self.mouse.move_to(self.win.game_view.get_center())
+        pag.keyDown('up')
+        time.sleep(2)
+        pag.keyUp('up')
+        time.sleep(0.5)
+
+    def set_camera_zoom(self, percentage: int) -> bool:
+        '''
+        Sets the camera zoom level.
+        Args:
+            percentage: The percentage of the camera zoom level to set.
+        Returns:
+            True if the zoom level was set, False if an issue occured.
+        '''
+        if percentage < 1:
+            percentage = 1
+        elif percentage > 100:
+            percentage = 100
+        self.log_msg(f"Setting camera zoom to {percentage}%...")
+        if not self.__open_display_settings():
+            return False
+        scroll_rect = Rectangle(left=self.win.control_panel.left + 84, top=self.win.control_panel.top + 146,
+                                width=102, height=8)
+        x = int((percentage / 100) * (scroll_rect.left + scroll_rect.width - scroll_rect.left) + scroll_rect.left)
+        p = scroll_rect.random_point()
+        self.mouse.move_to((x, p.y))
+        self.mouse.click()
+        return True
+    
+    def __open_display_settings(self) -> bool:
+        '''
+        Opens the display settings for the game client.
+        Returns:
+            True if the settings were opened, False if an error occured.
+        '''
+        control_panel = self.win.control_panel
+        self.mouse.move_to(self.win.cp_tabs[11].random_point())
+        self.mouse.click()
+        time.sleep(0.5)
+        display_tab = bcv.search_img_in_rect(f"{bcv.BOT_IMAGES}/cp_settings_display_tab.png", control_panel)
+        if display_tab is None:
+            self.log_msg("Could not find the display settings tab.")
+            return False
+        self.mouse.move_to(display_tab.random_point())
+        self.mouse.click()
+        time.sleep(0.5)
+        return True
